+++ cmd/eksctl-anywhere/cmd/installpackagecontroller.go
@@ -63,7 +63,7 @@ func installPackageController(ctx context.Context) error {
 
 	ctrlClient := deps.PackageControllerClient
 
-	if ctrlClient.IsInstalled(ctx) {
+	if clusterSpec.Cluster.IsSelfManaged() && ctrlClient.IsInstalled(ctx) {
 		return errors.New("curated Packages controller exists in the current cluster")
 	}